@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:moengage_cards/moengage_cards.dart' as moe;
-import 'package:moengage_flutter_example/cards/card_widget.dart';
+import 'card_widget.dart';
 
 class CardsScreen extends StatefulWidget {
-  const CardsScreen({Key? key}) : super(key: key);
+  const CardsScreen({super.key});
 
   @override
   State<CardsScreen> createState() => _CardsScreenState();
@@ -13,7 +13,7 @@ class _CardsScreenState extends State<CardsScreen>
     with TickerProviderStateMixin {
   List<moe.Card> cardList = [];
 
-  moe.MoEngageCards cards = moe.MoEngageCards("DAO6UGZ73D9RTK8B5W96TPYN");
+  moe.MoEngageCards cards = moe.MoEngageCards('DAO6UGZ73D9RTK8B5W96TPYN');
 
   List<String> categories = [];
 
@@ -24,10 +24,10 @@ class _CardsScreenState extends State<CardsScreen>
   @override
   void initState() {
     super.initState();
-    cards.setAppOpenCardsSyncListener((data) {
-      debugPrint("Cards App Open Sync Listener: $data");
+    cards.setAppOpenCardsSyncListener((moe.SyncCompleteData? data) {
+      debugPrint('Cards App Open Sync Listener: $data');
     });
-    cards.onCardsSectionLoaded((data) {
+    cards.onCardsSectionLoaded((moe.SyncCompleteData? data) {
       if (data?.hasUpdates == true) {
         setState(() {
           showHasUpdates = true;
@@ -36,7 +36,6 @@ class _CardsScreenState extends State<CardsScreen>
     });
     fetchCards();
     tabController = TabController(
-      initialIndex: 0,
       length: categories.length,
       vsync: this,
     );
@@ -44,16 +43,15 @@ class _CardsScreenState extends State<CardsScreen>
   }
 
   fetchCards() {
-    cards.getCardsInfo().then((data) {
+    cards.getCardsInfo().then((moe.CardsInfo data) {
       setState(() {
         cardList = data.cards;
         categories.clear();
-        if (data.shouldShowAllTab && data.categories.length > 0) {
-          categories.add("All");
+        if (data.shouldShowAllTab && data.categories.isNotEmpty) {
+          categories.add('All');
         }
         categories.addAll(data.categories);
         tabController = TabController(
-          initialIndex: 0,
           length: categories.length,
           vsync: this,
         );
@@ -71,11 +69,11 @@ class _CardsScreenState extends State<CardsScreen>
         child: Scaffold(
           appBar: AppBar(
             centerTitle: true,
-            iconTheme: IconThemeData(color: Colors.black54),
+            iconTheme: const IconThemeData(color: Colors.black54),
             backgroundColor: Colors.white,
             shadowColor: Colors.grey.shade100,
-            title: Text(
-              "Inbox",
+            title: const Text(
+              'Inbox',
               style: TextStyle(color: Colors.black54),
             ),
           ),
@@ -84,14 +82,13 @@ class _CardsScreenState extends State<CardsScreen>
                 refreshCards();
               },
               child: SingleChildScrollView(
-                scrollDirection: Axis.vertical,
-                physics: AlwaysScrollableScrollPhysics(),
-                child: Container(
+                physics: const AlwaysScrollableScrollPhysics(),
+                child: SizedBox(
                   height: MediaQuery.of(context).size.height -
                       AppBar().preferredSize.height,
                   child: (categories.isEmpty)
-                      ? Center(
-                          child: Text("No Data"),
+                      ? const Center(
+                          child: Text('No Data'),
                         )
                       : getListWidget(),
                 ),
@@ -109,18 +106,18 @@ class _CardsScreenState extends State<CardsScreen>
               children: [
                 TabBar(
                   isScrollable: true,
-                  padding: EdgeInsets.all(8.0),
-                  indicator: UnderlineTabIndicator(
+                  padding: const EdgeInsets.all(8.0),
+                  indicator: const UnderlineTabIndicator(
                     borderSide: BorderSide(width: 3.0, color: Colors.black54),
                     insets: EdgeInsets.symmetric(horizontal: 16.0),
                   ),
                   indicatorSize: TabBarIndicatorSize.label,
                   tabs: categories
-                      .map((category) => Padding(
-                            padding: EdgeInsets.symmetric(vertical: 8),
+                      .map((String category) => Padding(
+                            padding: const EdgeInsets.symmetric(vertical: 8),
                             child: Text(
                               category,
-                              style: TextStyle(
+                              style: const TextStyle(
                                 color: Colors.black38,
                                 fontSize: 18,
                               ),
@@ -133,24 +130,29 @@ class _CardsScreenState extends State<CardsScreen>
                   child: TabBarView(
                       controller: tabController,
                       children: categories
-                          .map((category) => getCardsByCategory(category))
-                          .map((list) {
+                          .map(
+                              (String category) => getCardsByCategory(category))
+                          .map((Future<List<moe.Card>> list) {
                         return Container(
                           child: FutureBuilder<List<moe.Card>>(
                               future: list,
-                              builder: (context, snapshot) {
+                              builder: (BuildContext context,
+                                  AsyncSnapshot<List<moe.Card>> snapshot) {
                                 if (snapshot.hasData &&
                                     snapshot.connectionState ==
                                         ConnectionState.done) {
-                                  final data = snapshot.data ?? [];
+                                  final List<moe.Card> data =
+                                      snapshot.data ?? [];
                                   if (data.isEmpty) {
-                                    return Center(child: Text("No Data"));
+                                    return const Center(child: Text('No Data'));
                                   }
                                   return ListView.builder(
                                       itemCount: data.length,
                                       shrinkWrap: true,
-                                      physics: AlwaysScrollableScrollPhysics(),
-                                      itemBuilder: (context, pos) {
+                                      physics:
+                                          const AlwaysScrollableScrollPhysics(),
+                                      itemBuilder:
+                                          (BuildContext context, int pos) {
                                         return (data[pos]
                                                     .template
                                                     .templateType ==
@@ -161,7 +163,7 @@ class _CardsScreenState extends State<CardsScreen>
                                                 data[pos], actionCallback);
                                       });
                                 }
-                                return Center(
+                                return const Center(
                                     child: CircularProgressIndicator());
                               }),
                         );
@@ -176,37 +178,33 @@ class _CardsScreenState extends State<CardsScreen>
                   Padding(
                     padding: const EdgeInsets.only(top: 24.0),
                     child: OutlinedButton(
-                      child: Text(
-                        "New Updates",
-                        style: TextStyle(color: Colors.black54),
-                      ),
                       onPressed: () {
                         fetchCards();
                         setState(() {
-                          this.showHasUpdates = false;
+                          showHasUpdates = false;
                         });
                       },
                       style: ButtonStyle(
                         backgroundColor:
                             MaterialStateProperty.all(Colors.white),
-                        shape:
-                            MaterialStateProperty.all<RoundedRectangleBorder>(
-                                RoundedRectangleBorder(
-                                    borderRadius: BorderRadius.circular(18.0),
-                                    side: BorderSide(color: Colors.black54))),
+                        shape: MaterialStateProperty
+                            .all<RoundedRectangleBorder>(RoundedRectangleBorder(
+                                borderRadius: BorderRadius.circular(18.0),
+                                side: const BorderSide(color: Colors.black54))),
                         side: MaterialStateProperty.all(
-                          BorderSide(
-                              color: Colors.black,
-                              width: 1.0,
-                              style: BorderStyle.solid),
+                          const BorderSide(),
                         ),
+                      ),
+                      child: const Text(
+                        'New Updates',
+                        style: TextStyle(color: Colors.black54),
                       ),
                     ),
                   ),
                 ],
               )
             else
-              SizedBox.shrink()
+              const SizedBox.shrink()
           ],
         )
       ],
@@ -214,7 +212,7 @@ class _CardsScreenState extends State<CardsScreen>
   }
 
   void refreshCards() {
-    cards.refreshCards((data) {
+    cards.refreshCards((moe.SyncCompleteData? data) {
       cards.cardDelivered();
       if (data?.hasUpdates == true) {
         fetchCards();
@@ -248,8 +246,8 @@ class _CardsScreenState extends State<CardsScreen>
   }
 
   Future<List<moe.Card>> getCardsByCategory(String category) async {
-    if (category == "All") return cardList;
-    final cardData = await cards.getCardsForCategory(category);
+    if (category == 'All') return cardList;
+    final moe.CardsData cardData = await cards.getCardsForCategory(category);
     return cardData.cards;
   }
 }
